+++ python_plus/tests/python_plus_test_01.py
@@ -18,7 +18,7 @@ MODULE_ID = 'python_plus'
 TEST_FAILED = 1
 TEST_SUCCESS = 0
 
-__version__ = "1.0.2"
+__version__ = "1.0.2.1"
 
 
 def version():