+++ wok_code/manage_pypi.sh
@@ -1,5 +1,5 @@
 # set -x
-__version__=1.0.1.21
+__version__=1.0.1.22
 act=""
 pypi=""
 opts=""